+++ src/AngleSharp/Html/Parser/HtmlTreeMode.cs
@@ -1,4 +1,4 @@
-﻿namespace AngleSharp.Html.Parser
+namespace AngleSharp.Html.Parser
 {
     /// <summary>
     /// Possible insertation mode values.
@@ -92,6 +92,6 @@
         /// <summary>
         /// Once we are far behind the frameset tag.
         /// </summary>
-        AfterAfterFrameset
+        AfterAfterFrameset,
     }
 }